Handle Spring 4.3.16 Bean Creation Regression Spring 4.3.16+ now loads outer classes when determining appropriate constructor parameters for static inner classes. Since these classes were manually registered as beans and they also were within classes whose hierarchies dependend on TestNG, this caused all integration tests to fail when TestNG was not on the classpath.
This simply moves those classes out to their own files so they can exist without any TestNG dependencies
