+++ integration/src/test/java/org/broadleafcommerce/common/workflow/DummyProcessContextFactory.java
@@ -1,0 +1,34 @@
+/*
+ * #%L
+ * BroadleafCommerce Integration
+ * %%
+ * Copyright (C) 2009 - 2018 Broadleaf Commerce
+ * %%
+ * Licensed under the Broadleaf Fair Use License Agreement, Version 1.0
+ * (the "Fair Use License" located  at http://license.broadleafcommerce.org/fair_use_license-1.0.txt)
+ * unless the restrictions on use therein are violated and require payment to Broadleaf in which case
+ * the Broadleaf End User License Agreement (EULA), Version 1.1
+ * (the "Commercial License" located at http://license.broadleafcommerce.org/commercial_license-1.1.txt)
+ * shall apply.
+ * 
+ * Alternatively, the Commercial License may be replaced with a mutually agreed upon license (the "Custom License")
+ * between you and Broadleaf Commerce. You may not use this file except in compliance with the applicable license.
+ * #L%
+ */
+package org.broadleafcommerce.common.workflow;
+
+import org.broadleafcommerce.core.workflow.DefaultProcessContextImpl;
+import org.broadleafcommerce.core.workflow.ProcessContext;
+import org.broadleafcommerce.core.workflow.ProcessContextFactory;
+import org.broadleafcommerce.core.workflow.WorkflowException;
+
+public class DummyProcessContextFactory implements ProcessContextFactory<Object, Object> {
+
+    @Override
+    public ProcessContext<Object> createContext(Object preSeedData) throws WorkflowException {
+        ProcessContext<Object> context = new DefaultProcessContextImpl<>();
+        context.setSeedData(preSeedData);
+        return context;
+    }
+    
+}
