+++ integration/src/test/java/org/broadleafcommerce/common/workflow/SimpleRollbackHandler.java
@@ -1,0 +1,35 @@
+/*
+ * #%L
+ * BroadleafCommerce Integration
+ * %%
+ * Copyright (C) 2009 - 2018 Broadleaf Commerce
+ * %%
+ * Licensed under the Broadleaf Fair Use License Agreement, Version 1.0
+ * (the "Fair Use License" located  at http://license.broadleafcommerce.org/fair_use_license-1.0.txt)
+ * unless the restrictions on use therein are violated and require payment to Broadleaf in which case
+ * the Broadleaf End User License Agreement (EULA), Version 1.1
+ * (the "Commercial License" located at http://license.broadleafcommerce.org/commercial_license-1.1.txt)
+ * shall apply.
+ * 
+ * Alternatively, the Commercial License may be replaced with a mutually agreed upon license (the "Custom License")
+ * between you and Broadleaf Commerce. You may not use this file except in compliance with the applicable license.
+ * #L%
+ */
+package org.broadleafcommerce.common.workflow;
+
+import org.broadleafcommerce.core.workflow.Activity;
+import org.broadleafcommerce.core.workflow.ProcessContext;
+import org.broadleafcommerce.core.workflow.state.RollbackFailureException;
+import org.broadleafcommerce.core.workflow.state.RollbackHandler;
+
+import java.util.List;
+import java.util.Map;
+
+public class SimpleRollbackHandler implements RollbackHandler<ProcessContext<List<String>>> {
+
+    @Override
+    public void rollbackState(Activity<ProcessContext<List<String>>> activity, ProcessContext<List<String>> processContext, Map<String, Object> stateConfiguration) throws RollbackFailureException {
+        processContext.getSeedData().add("Rollback" + ((SimpleActivity) activity).getName());
+    }
+    
+}
