+++ integration/src/test/resources/bl-applicationContext-test.xml
@@ -116,37 +116,37 @@
     
     <bean id="testRollbackWorkflow" class="org.broadleafcommerce.core.workflow.SequenceProcessor">
         <property name="processContextFactory">
-            <bean class="org.broadleafcommerce.common.workflow.RollbackTest.DummyProcessContextFactory"/>
+            <bean class="org.broadleafcommerce.common.workflow.DummyProcessContextFactory"/>
         </property>
         <property name="activities">
             <list>
-                <bean class="org.broadleafcommerce.common.workflow.RollbackTest.SimpleActivity">
+                <bean class="org.broadleafcommerce.common.workflow.SimpleActivity">
                     <constructor-arg value="Activity1"></constructor-arg>
                     <property name="rollbackHandler">
-                        <bean class="org.broadleafcommerce.common.workflow.RollbackTest.SimpleRollbackHandler" />
+                        <bean class="org.broadleafcommerce.common.workflow.SimpleRollbackHandler" />
                     </property>
                 </bean>
-                <bean class="org.broadleafcommerce.common.workflow.RollbackTest.SimpleActivity">
+                <bean class="org.broadleafcommerce.common.workflow.SimpleActivity">
                     <constructor-arg value="Activity2"></constructor-arg>
                     <property name="rollbackHandler">
-                        <bean class="org.broadleafcommerce.common.workflow.RollbackTest.SimpleRollbackHandler" />
+                        <bean class="org.broadleafcommerce.common.workflow.SimpleRollbackHandler" />
                     </property>
                 </bean>
-                <bean class="org.broadleafcommerce.common.workflow.RollbackTest.NestedActivity">
+                <bean class="org.broadleafcommerce.common.workflow.NestedActivity">
                     <constructor-arg name="workflow">
                         <bean class="org.broadleafcommerce.core.workflow.SequenceProcessor">
                             <property name="processContextFactory">
-                                <bean class="org.broadleafcommerce.common.workflow.RollbackTest.DummyProcessContextFactory"/>
+                                <bean class="org.broadleafcommerce.common.workflow.DummyProcessContextFactory"/>
                             </property>
                             <property name="activities">
                                 <list>
-                                    <bean class="org.broadleafcommerce.common.workflow.RollbackTest.SimpleActivity">
+                                    <bean class="org.broadleafcommerce.common.workflow.SimpleActivity">
                                         <constructor-arg value="ActivityA"></constructor-arg>
                                         <property name="rollbackHandler">
-                                            <bean class="org.broadleafcommerce.common.workflow.RollbackTest.SimpleRollbackHandler" />
+                                            <bean class="org.broadleafcommerce.common.workflow.SimpleRollbackHandler" />
                                         </property>
                                     </bean>
-                                    <bean class="org.broadleafcommerce.common.workflow.RollbackTest.ExceptionActivity" />
+                                    <bean class="org.broadleafcommerce.common.workflow.ExceptionActivity" />
                                 </list>
                             </property>
                             <property name="defaultErrorHandler" ref="blSilentErrorHandler"/>
